tests/requirements-test.txt
pytest==8.3.0
pytest-playwright==0.5.1
pytest-bdd==7.3.0
python-dotenv==1.0.1
